+++ scene/r_bump_checker.rt
@@ -12,10 +12,10 @@ L       5,5,-15			0.5				255,255,255
 
 # Sphere
 ###     xyx             diameter        rgb				mat & tex		bump		filename			
-sp      -15,0,0         20              255,0,0			metal			ON			NONE
+# sp      -15,0,0         20              255,0,0			metal			ON			NONE
 sp      15,0,0	         20             0,255,0			checker			OFF			NONE
 
 # # Plane
 # ### 	xyz             vector          rgb	
 pl      0,-20,0       	0,1,0           127,127,127		metal			ON			NONE
-pl		0,0,40			0,0,1			255,255,255		OFF				OFF			NONE
+pl		0,0,80			0,0,1			255,255,255		OFF				OFF			NONE
