+++ src/raytracing/raytracing_utils.c
